+++ docs/07-Templates/list-page.mdx
@@ -13,7 +13,7 @@ import ListPage from '@site/src/components/templates/ListPage/ListPage';
 import { DataProvider } from '@site/src/components/templates/ListPage/DataProvider';
 
 import ListPageRaw from '!!raw-loader!@site/src/components/templates/ListPage/ListPage';
-import types from '!!raw-loader!@site/src/components/templates/CardListPage/types';
+import types from '!!raw-loader!@site/src/components/templates/GridPage/types';
 
 ## Overview
