+++ src/components/templates/GridPage/types.ts
